mysql적용

+++ controller/post.mjs
@@ -26,8 +26,9 @@ export async function getPostId(req, res, next) {
 
 // 포스트를 생성하는 함수
 export async function createPost(req, res, next) {
-  const { userid, name, text } = req.body;
-  const posts = await postRepository.create(userid, name, text);
+  const  {text}  = req.body;
+  console.log("req.useridx:", req.useridx)
+  const posts = await postRepository.create(text,req.useridx);
   res.status(201).json(posts);
 }
 
@@ -47,6 +48,13 @@ export async function updatePost(req, res, next) {
 // 포스트 삭제하는 함수
 export async function deletePost(req, res, next) {
   const id = req.params.id;
-  const posts = await postRepository.remove(id);
-  res.status(200).json(posts);
+  const post = await postRepository.getAllById(id);
+  if(!post){
+    return res.status(404).json({message : `${id}의 포스트가 없습니다.`})
+  }
+  if(post.useridx !== req.useridx){
+    return res.sendStatus(403);
+  }
+  await postRepository.remove(id);
+  res.sendStatus(204);
 }
